+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 1.0.0
+# 0.1.0
 
-- Initial version.
+- Initial release of `m3u_xmltv`.
+- Added stream and file parsing for M3U playlists.
+- Added stream parser for XMLTV EPG data.
+- Added `EpgMatcher` for mapping M3U channels with XMLTV programmes.